Positioning, bluetooth and lidar works together and sends the data over REST API
